feature/sinusoidal-walking - restructured folder structure
